Add experimental tag disk sync
It's very simplistic for now, just detects any filesystem changes and calls tag reload if it thinks it's a tag file. Didn't test what happens with tags that are not used, or are invalid. Also need to look into a better check for detecting when a new scenario is being loaded.