[Guerilla] fix crash when opening tags in documents
Crash was caused by some careless patching.

The toolkit usually incorrectly uses a wide path to the documents folder as a narrow path, to fix this h2codez patched the code to return a narrow path, however some code was correctly using the wide path. This was fixed by patching that code to call the original function that returned a wide path. The crash occurred as a result of not patching all instances of such code.